Moved door generation logic to own class
