+++ Assets/Scripts/Generators/RoomGenerator.cs
@@ -24,8 +24,6 @@ public class RoomGenerator : DungeonTaskBase {
             context.createdRooms[currentRoom.Center] = currentRoom;
             AddNeighbour(currentRoom, roomsToCreate);
         }
-
-        CreateDoors();
     }
 
     private void AddNeighbour(Room currentRoom, Queue<Room> roomsToCreate) {
@@ -64,15 +62,5 @@ public class RoomGenerator : DungeonTaskBase {
             .First();
 
         availableNeighbours.Remove(closest);
-    }
-
-    private void CreateDoors() {
-        foreach(Room room in context.createdRooms.Values) {
-            foreach(Vector2Int pos in room.GetNeighbourPositions()) {
-                if(context.createdRooms.TryGetValue(pos, out Room neighbour)) {
-                    room.Connect(neighbour);
-                }
-            }
-        }
     }
 }
